solve some warning of files

+++ source/common/davs2.cc
@@ -490,7 +490,7 @@ davs2_decoder_open(davs2_param_t *param)
 
     /* init log module */
     mgr->module_log.i_log_level = param->info_level;
-    sprintf(mgr->module_log.module_name, "Manager %06llx", (intptr_t)(mgr));
+    sprintf(mgr->module_log.module_name, "Manager %06llx", (long long unsigned int)(mgr));
 
     if (mgr->param.threads <= 0) {
         mgr->param.threads = davs2_cpu_num_processors();
